// src/models/common.ts
export interface FontSliceProps {
  isNoto?: boolean;
  subsets?: string[];
  category?: any;
  family?: string;
  data: any;
  loading: boolean;
  error: string | null;
}
export interface TesterDescription {
  data: string;
}
export interface RangeFontProps {
  name: string;
}
export interface MultiSelectProps {
  name: string;
  id: string | number;
}